requirements/requirements.txt
django==1.4.2
psycopg2

south

django-pagination
django_compressor
python-memcached

PIL
sorl-thumbnail

django-tastypie